simplify generated code hasValue

+++ plugins/org.chromium.sdk/src/org/chromium/sdk/internal/v8native/protocol/input/FrameObject.java
@@ -53,9 +53,9 @@ public interface FrameObject {
 
   String type();
 
-  Long position();
+  Integer position();
 
-  Long column();
+  Integer column();
 
   Boolean debuggerFrame();
 }
